+++ src/Constraints/Insee.php
@@ -2,15 +2,13 @@
 
 namespace SLLH\IsoCodesValidator\Constraints;
 
-use Symfony\Component\Validator\Constraint;
-
 /**
  * @Annotation
  * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
  *
  * @author Sullivan Senechal <soullivaneuh@gmail.com>
  */
-class Insee extends Constraint
+class Insee extends IsoCodesGeneric
 {
     public $message = 'This INSEE number is not valid.';
 }