+++ src/Constraints/Uknin.php
@@ -2,15 +2,13 @@
 
 namespace SLLH\IsoCodesValidator\Constraints;
 
-use Symfony\Component\Validator\Constraint;
-
 /**
  * @Annotation
  * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
  *
  * @author Sullivan Senechal <soullivaneuh@gmail.com>
  */
-class Uknin extends Constraint
+class Uknin extends IsoCodesGeneric
 {
     public $message = 'This value is not a valid NINO.';
 }